requirements.txt
openai
pandas
streamlit
python-dotenv
requests
